requirements.txt
streamlit
pandas
mysql-connector-python
plotly
statsmodels
python-dotenv
